Health updates for 1.6

+++ src/com/garbagemule/MobArena/health/HealthStrategyHeroes.java
@@ -13,8 +13,7 @@ public class HealthStrategyHeroes implements HealthStrategy
         double regain  = health == p.getMaxHealth() ? p.getMaxHealth() : health - current;
         
         try {
-            // TODO: Remove cast for 1.6
-            EntityRegainHealthEvent event = new EntityRegainHealthEvent(p, (int) regain, RegainReason.CUSTOM);
+            EntityRegainHealthEvent event = new EntityRegainHealthEvent(p, regain, RegainReason.CUSTOM);
             Bukkit.getPluginManager().callEvent(event);
         }
         catch (Exception e) {} // Because Bukkit is retarded.
